Sometimes Content-Type is not set
This is probably the final patch for issues #GH-10 and #GH-12

The body of the request isn't included in the signature base string if the
request's 'Content-Type' header isn't "application/x-www-form-urlencoded".

Also, requests-oauth overwrites request.data, relying on requests reenconding
correctly if `request.data` or `request.files` was modified in the hook. This
way we avoid duplicating code from requests and hooks can alter body to be sent.

This way we can also pull some parameters out of the data, such as oauth_callback.
This way we get #GH-13 working.