Fixing _enc_params and cleaning data
url used in requests is calculated different, if we reset _enc_params as we had to do before,
we can lose GET parameters.

Also if we are doing a POST request, if information gets encoded in the url, we clean data, for
avoiding sending duplicate information, which makes many OAuth providers fail.